requirements.txt
# Core dependencies for AdaptiveGraspControl Project (Stable Version)
# Stage 0: Environment Setup and Basic Validation

# Serial communication
pyserial==3.5
pyserial-asyncio==0.6

# Data processing and scientific computing
numpy==1.24.3
scipy==1.10.1
pandas==2.0.3

# Real-time data visualization
matplotlib==3.7.1
pyqtgraph==0.13.3

# GUI framework
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1

# Configuration file processing
pyyaml==6.0
configparser==5.3.0

# Logging and debugging
colorlog==6.7.0
python-dotenv==1.0.0

# Testing framework
pytest==7.4.0
pytest-cov==4.1.0
pytest-timeout==2.1.0

# Development tools (core only)
ipython==8.13.0
black==23.3.0
flake8==6.0.0
pylint==2.17.4

# Performance analysis
line-profiler==4.0.3
memory-profiler==0.61.0

# Documentation generation (compatible versions)
sphinx==6.2.1
sphinx-rtd-theme==1.2.2